# codec/types: avoid unnecessary allocations for NewAnyWithCustomTypeURL on error