@odeke-em 2021-02-17
Gotcha, thanks Marko!

On Wed, Feb 17, 2021 at 1:54 AM Marko <notifications@github.com> wrote:

> I kindly request that we backport this change too.
>
> on merge the backport should be handled unless conflicts arise.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/cosmos/cosmos-sdk/pull/8605#issuecomment-780438819>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABFL3VZCNTKOUQ65YNXFTY3S7OG4XANCNFSM4XYA6RLQ>
> .
>