@robert-zaremba 2021-02-17
Good job.

PS: I thought we backport only fixes and significant changes? This is a micro optimization of an _unhappy_ path
PS2: All optimizations are always good :clap: