@odeke-em 2021-02-17
Great question Robert! I asked for a backport as it isn’t just an
optimization but supposedly if someone could trivially send unmarshallable
data in a fast loop they could cause memory to be allocated unnecessarily,
which is why I called it a slow bleed — better for security to seal off
such seemingly simple paths but that can cause unnecessary allocations.

On Wed, Feb 17, 2021 at 5:32 AM Robert Zaremba <notifications@github.com>
wrote:

> Good job.
>
> PS: I thought we backport only fixes and significant changes? This is a
> micro optimization of an *unhappy* path
> PS2: All optimizations are always good 👏
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/cosmos/cosmos-sdk/pull/8605#issuecomment-780557936>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABFL3VZ2AJNWINUGZHSNO2TS7PAOVANCNFSM4XYA6RLQ>
> .
>